websockets, django channels, frontend updates

+++ songrequestapp/views.py
@@ -1,3 +1,4 @@
+from email import message
 from django.shortcuts import render, redirect, get_object_or_404
 from django.contrib.auth import login
 from django.views import View
@@ -10,9 +11,10 @@ from django.contrib.auth.decorators import login_required, user_passes_test
 from pymongo import MongoClient
 from datetime import datetime
 from datetime import timedelta
-from .models import song
+from .models import song, songs_play_history
 from .youtube_api import scrape_title, video_id, get_sec
 from bson.objectid import ObjectId
+from .consumers import WSConsumer
 # Create your views here.
 
 client = MongoClient("mongodb+srv://shiro_47:cJOKjP8VgVrl0M7l@song-request.la9qn.mongodb.net/?retryWrites=true&w=majority")
@@ -40,23 +42,43 @@ def request_song(request):
         songs_blacklist=db["songs_blacklist"]
         try:
             if song.objects.get(song_url=request.POST["YouTube URL"]):
+                print(request.POST["YouTube URL"])
                 return JsonResponse({"message":"ALREADY IN QUEUE"})
         except ObjectDoesNotExist:
-            if songs_blacklist.find_one({"url": request.POST["YouTube URL"]})!=None or songs_blacklist.find_one({"title": scrape_title(request.POST["YouTube URL"])[0]})!=None:
-                return JsonResponse({"message": "SONG IS BLACKLISTED"})
             if "www.youtube.com" in request.POST["YouTube URL"] or "youtu.be" in request.POST["YouTube URL"]:
-                url= request.POST["YouTube URL"]
-                data=scrape_title(url)
-                Song=song()
-                Song.song_url=url
-                Song.song_title=data[0]
-                Song.song_duration= data[1]
-                Song.song_thumbnail= data[2]
+                url = request.POST["YouTube URL"]
+                data = scrape_title(url)
+                if data==None:
+                    return JsonResponse({"message": "INVALID URL"})
+                if songs_blacklist.find_one({"url": url}) != None or songs_blacklist.find_one({"title": data[0]}) != None:
+                    return JsonResponse({"message": "SONG IS BLACKLISTED"})
+                Song = song()
+                Song.song_url = url
+                Song.song_title = data[0]
+                Song.song_duration = data[1]
+                Song.song_thumbnail = data[2]
                 Song.save()
-                return JsonResponse({"message":"SONG REQUESTED"})
+                return JsonResponse({"message": "SONG REQUESTED"})
             else:
-                return JsonResponse({"message":"INVALID URL"})
-    return render(request, 'request_song.html')
+                return JsonResponse({"message": "INVALID URL"})
+    try:
+        songs= song.objects.all()
+        current_time = datetime.now()
+        sec=0
+        songs_play_times=[]
+        for x in songs:
+            if x==songs[0]:
+                songs_play_times.append(current_time.strftime('%H:%M:%S'))
+            time1= x.song_duration
+            sec+= get_sec(time1)
+            estimated_time=current_time+timedelta(seconds=sec)
+            songs_play_times.append(estimated_time.strftime('%H:%M:%S'))
+        elements={
+            "songs": zip(songs,songs_play_times),
+        }
+        return render(request, 'request_song.html', elements)
+    except: 
+        return render(request, 'request_song.html')
 
 
 @login_required(login_url='/accounts/login/')
@@ -64,25 +86,22 @@ def request_song(request):
 def dashboard(request):
     try:
         songs= song.objects.all()
-        current_song_title=songs[0].song_title
-        current_song_url=songs[0].song_url
         current_time = datetime.now()
         sec=0
         songs_play_times=[]
         for x in songs:
             if x==songs[0]:
                 songs_play_times.append(current_time.strftime('%H:%M:%S'))
-                time1= x.song_duration
-                sec+= get_sec(time1)
             time1= x.song_duration
             sec+= get_sec(time1)
             estimated_time=current_time+timedelta(seconds=sec)
             songs_play_times.append(estimated_time.strftime('%H:%M:%S'))
         dict={"songs":zip(songs,songs_play_times),
-            "current_song_url": current_song_url,
-            "current_song_title": current_song_title,
+            "current_song_url": video_id(songs[0].song_url),
+            "current_song_title": songs[0].song_title,
             "current_song_id": songs[0].id,
-            "song_id":video_id(current_song_url),
+            "current_song_skip_requests": songs[0].skip_requests,
+            "song_id":video_id(songs[0].song_url),
                     }
         return render(request, 'dashboard.html', dict)
     except:
@@ -98,10 +117,32 @@ def songs_blacklist(request):
     return render(request, "songs_blacklist.html", dict)
 
 @user_passes_test(lambda u: u.is_superuser)
+def songs_history(request):
+    try:
+        songs= songs_play_history.objects.all()
+        dict={"songs":songs,
+        }
+        return render(request, 'history.html', dict)
+    except:
+        return render(request, 'history.html')
+
+@user_passes_test(lambda u: u.is_superuser)
 def next_song(request):
     try:
+        skip_votes=db["skip_votes"]
         songs= song.objects.all()
+        Song=songs_play_history()
+        Song.song_url = songs[0].song_url
+        Song.song_title = songs[0].song_title
+        Song.song_duration = songs[0].song_duration
+        Song.song_thumbnail = songs[0].song_thumbnail
+        Song.save()
         songs[0].delete()
+        skip_votes.delete_many({})
+        current_song=songs[0]
+        response={"current_song":{"title":current_song.song_title,
+                                  "url":current_song.song_url,
+                                  "id":current_song.id}}
         return redirect(reverse("songrequestapp:dashboard"))
     except IndexError:
         return render(request, 'dashboard.html')
@@ -114,6 +155,14 @@ def song_delete(request, id):
         song_that_is_ready_to_be_deleted.delete()
 
     return HttpResponseRedirect('/dashboard')
+
+@user_passes_test(lambda u: u.is_superuser)
+def song_delete_from_history(request, id):
+    song_that_is_ready_to_be_deleted = get_object_or_404(songs_play_history, id=id)
+    if request.method == 'POST':
+        song_that_is_ready_to_be_deleted.delete()
+
+    return HttpResponseRedirect('/dashboard/history')
 
 @user_passes_test(lambda u: u.is_superuser)
 def add_song_to_blacklist(request, id): 
@@ -129,8 +178,21 @@ def add_song_to_blacklist(request, id):
 
 @user_passes_test(lambda u: u.is_superuser)
 def remove_from_blacklist(request, id): 
-    print('fsd')
-    songs_blacklist=db["songs_blacklist"]
-    songs_blacklist.delete_one({"_id": ObjectId(id)})
-    return JsonResponse({"message":"DELETED"})
+    if request.method == "POST":
+        print('fsd')
+        songs_blacklist=db["songs_blacklist"]
+        songs_blacklist.delete_one({"_id": ObjectId(id)})
+        return JsonResponse({"message":"DELETED"})
 
+@login_required(login_url='/accounts/login/')
+def skip_vote(request, id): 
+    if request.method == "POST":
+        skip_votes=db["skip_votes"]
+        if skip_votes.find_one({"username": request.user.username}) != None:
+            return JsonResponse({"message":"ALREADY VOTED"})
+        user={"username": request.user.username}
+        # skip_votes.insert_one(user)
+        Song = get_object_or_404(song, id=id)
+        Song.skip_requests +=1
+        Song.save()
+        return JsonResponse({"message":"VOTED"})
